+++ lib/widgets/habits/habit_completion_card.dart
@@ -84,14 +84,15 @@ class _HabitCompletionCardState extends State<HabitCompletionCard> {
                       Row(
                         children: [
                           Visibility(
-                              visible: widget.habitDefinition.priority ?? false,
-                              child: Padding(
-                                padding: const EdgeInsets.only(right: 5),
-                                child: Icon(
-                                  Icons.star,
-                                  color: styleConfig().starredGold,
-                                ),
-                              )),
+                            visible: widget.habitDefinition.priority ?? false,
+                            child: Padding(
+                              padding: const EdgeInsets.only(right: 5),
+                              child: Icon(
+                                Icons.star,
+                                color: styleConfig().starredGold,
+                              ),
+                            ),
+                          ),
                           Text(
                             widget.habitDefinition.name,
                             style: completedToday